requirements.txt
requests
python-dotenv
typer
distro
litellm
beautifulsoup4
newspaper3k
playwright
lxml_html_clean
PyQt5
markdown2
reportlab
python-docx
instructor
sentence-transformers
torch
numpy
spacy
nltk
scikit-learn
fuzzywuzzy
python-levenshtein
